backend/requirements.txt
django~=4.2.3
djangorestframework~=3.14.0
redis~=4.6
rq~=1.15.1
```